refactor: rename namespace 'Tasks\*' to 'Task\*' and corresponding folder structure
