+++ src/tests/actions/CancelActionTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use R794021\Action\CancelAction;
-use R794021\Tasks\Task;
+use R794021\Task\Task;
 use R794021\Users\{Contractor, Customer};
 
 $cancel = new CancelAction();
