+++ src/tests/actions/DoneActionTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use R794021\Action\DoneAction;
-use R794021\Tasks\Task;
+use R794021\Task\Task;
 use R794021\Users\{Contractor, Customer};
 
 $done = new DoneAction();
