+++ src/tests/tasks/TaskTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use R794021\Tasks\Task;
+use R794021\Task\Task;
 use R794021\Users\{Contractor, Customer};
 use R794021\Action\{ApplyAction, CancelAction, DoneAction, RejectAction};
